moves the navigation item menu to a standalone file
