+++ packages/block-library/src/navigation/edit/menu-inspector-controls.js
@@ -1,7 +1,6 @@
 /**
  * WordPress dependencies
  */
-import { createBlock } from '@wordpress/blocks';
 import {
 	__experimentalOffCanvasEditor as OffCanvasEditor,
 	InspectorControls,
@@ -11,85 +10,16 @@ import {
 	PanelBody,
 	__experimentalHStack as HStack,
 	__experimentalHeading as Heading,
-	DropdownMenu,
-	MenuItem,
-	MenuGroup,
 } from '@wordpress/components';
-import { useSelect, useDispatch } from '@wordpress/data';
+import { useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
-import { moreVertical } from '@wordpress/icons';
 
 /**
  * Internal dependencies
  */
 import ManageMenusButton from './manage-menus-button';
 import NavigationMenuSelector from './navigation-menu-selector';
-
-const POPOVER_PROPS = {
-	className: 'block-editor-block-settings-menu__popover',
-	position: 'bottom right',
-	variant: 'toolbar',
-};
-
-const LeafMoreMenu = ( props ) => {
-	const { clientId, block } = props;
-
-	const { insertBlock, replaceBlock, removeBlocks } =
-		useDispatch( blockEditorStore );
-
-	return (
-		<DropdownMenu
-			icon={ moreVertical }
-			label={ __( 'Options' ) }
-			className="block-editor-block-settings-menu"
-			popoverProps={ POPOVER_PROPS }
-			noIcons
-			{ ...props }
-		>
-			{ ( { onClose } ) => (
-				<MenuGroup>
-					<MenuItem
-						onClick={ () => {
-							const newLink = createBlock(
-								'core/navigation-link'
-							);
-							if ( block.name === 'core/navigation-submenu' ) {
-								const updateSelectionOnInsert = false;
-								insertBlock(
-									newLink,
-									block.innerBlocks.length,
-									clientId,
-									updateSelectionOnInsert
-								);
-							} else {
-								// Convert to a submenu if the block currently isn't one.
-								const newSubmenu = createBlock(
-									'core/navigation-submenu',
-									block.attributes,
-									block.innerBlocks
-										? [ ...block.innerBlocks, newLink ]
-										: [ newLink ]
-								);
-								replaceBlock( clientId, newSubmenu );
-							}
-							onClose();
-						} }
-					>
-						{ __( 'Add submenu item' ) }
-					</MenuItem>
-					<MenuItem
-						onClick={ () => {
-							removeBlocks( [ clientId ], false );
-							onClose();
-						} }
-					>
-						{ __( 'Remove item' ) }
-					</MenuItem>
-				</MenuGroup>
-			) }
-		</DropdownMenu>
-	);
-};
+import { LeafMoreMenu } from '../leaf-more-menu';
 
 const MenuInspectorControls = ( {
 	clientId,
